testing and passing to about 78%
added circleci config.yml file
and remote db

added date on comments and message

added coveralls stuff

Delete src/tests directory

testing and passing to about 78%

added circleci config.yml file
and remote db

added date on comments and message

added coveralls stuff
